rename: 기본 엔티티 클래스가 있는 패키지명 도메인으로 수정
